Rename source files to fit naming convention
